refactor(ndt7): move results into its own package
