+++ ndt7/server/results/file.go
@@ -1,25 +1,27 @@
-package ndt7
+package results
 
 import (
 	"compress/gzip"
 	"encoding/json"
+	"net/http"
 	"os"
 	"time"
 
+	"github.com/gorilla/websocket"
 	"github.com/m-lab/ndt-cloud/ndt7/model"
 )
 
-// resultsfile is the file where we save measurements.
-type resultsfile struct {
+// File is the file where we save measurements.
+type File struct {
 	// Writer is the gzip writer instance
 	Writer *gzip.Writer
 	// Fp is the underlying file
 	Fp     *os.File
 }
 
-// newResultsfile opens a measurements file in the current working
+// newFile opens a measurements file in the current working
 // directory on success and returns an error on failure.
-func newResultsfile() (*resultsfile, error) {
+func newFile() (*File, error) {
 	format := "2006-01-02T15:04:05.000000000Z"
 	timestamp := time.Now().UTC().Format(format)
 	name := "ndt7-" + timestamp + ".jsonl.gz"
@@ -34,14 +36,35 @@ func newResultsfile() (*resultsfile, error) {
 		fp.Close()
 		return nil, err
 	}
-	return &resultsfile{
+	return &File{
 		Writer: writer,
 		Fp: fp,
 	}, nil
 }
 
+// OpenFor opens the results file and writes into it the results metadata based
+// on the query string. Returns the results file on success. Returns an error
+// in case of failure. The request argument is used to get the query string
+// containing the metadata. The conn argument is used to retrieve the local and
+// the remote endpoints addresses. The what argument should indicate whether
+// this is a "download" or an "upload" ndt7 measurement.
+func OpenFor(request *http.Request, conn *websocket.Conn, what string) (*File, error) {
+	meta := make(metadata)
+	initMetadata(&meta, conn.LocalAddr().String(),
+		conn.RemoteAddr().String(), request.URL.Query(), "download")
+	resultfp, err := newFile()
+	if err != nil {
+		return nil, err
+	}
+	if err := resultfp.WriteMetadata(meta); err != nil {
+		resultfp.Close()
+		return nil, err
+	}
+	return resultfp, nil
+}
+
 // Close closes the measurement file.
-func (fp *resultsfile) Close() error {
+func (fp *File) Close() error {
 	err := fp.Writer.Close()
 	if err != nil {
 		fp.Fp.Close()
@@ -60,7 +83,7 @@ type savedMeasurement struct {
 }
 
 // WriteMeasurement writes |measurement| on the measurements file.
-func (fp *resultsfile) WriteMeasurement(measurement model.Measurement, origin string) error {
+func (fp *File) WriteMeasurement(measurement model.Measurement, origin string) error {
 	return fp.writeInterface(savedMeasurement{
 		Origin: origin,
 		Measurement: measurement,
@@ -68,12 +91,12 @@ func (fp *resultsfile) WriteMeasurement(measurement model.Measurement, origin st
 }
 
 // WriteMetadata writes |metadata| on the measurements file.
-func (fp *resultsfile) WriteMetadata(metadata metadata) error {
+func (fp *File) WriteMetadata(metadata metadata) error {
 	return fp.writeInterface(metadata)
 }
 
 // writeInterface serializes |entry| as JSONL.
-func (fp *resultsfile) writeInterface(entry interface{}) error {
+func (fp *File) writeInterface(entry interface{}) error {
 	data, err := json.Marshal(entry)
 	if err != nil {
 		return err
